+++ docs/CHANGELOG.md
@@ -16,6 +16,7 @@
   * Names of parameters for `ConvGeneral` were standardized to "input," "kernel" and "channels."
 * Package `layers`:
   * Updated the configuration names for `Convolution`, to match the standards in the `graph` package.
+  * Added `ChannelGroupCount()` and `BatchGroupCount()` to `Convolution` configuration.
 * Updated to gopjrt v0.8.0, with the changes to the convolution API.
 
 # v0.21.1: 2025/08/16 Added Zero-dim tensors support and other small improvements.